+++ enums/CorsHeaderEnum.js
@@ -1,9 +1,17 @@
+/**
+ * Names of common CORS request headers.
+ */
 var CorsHeaderEnum;
 (function (CorsHeaderEnum) {
+    /** The `Accept` request header. */
     CorsHeaderEnum["Accept"] = "Accept";
+    /** The `Authorization` request header. */
     CorsHeaderEnum["Authorization"] = "Authorization";
+    /** The `Content-Type` request header. */
     CorsHeaderEnum["ContentType"] = "Content-Type";
+    /** The `Origin` request header. */
     CorsHeaderEnum["Origin"] = "Origin";
+    /** The `X-Requested-With` request header. */
     CorsHeaderEnum["XRequestedWith"] = "X-Requested-With";
 })(CorsHeaderEnum || (CorsHeaderEnum = {}));
 export default CorsHeaderEnum;