+++ facades/index.d.ts
@@ -1,1 +1,4 @@
-export * from "../facades/Cors";
+/**
+ * Re-exports the CORS facade.
+ */
+export { default as Cors } from "../facades/Cors";